requirements.txt
PyPDF2==3.0.1
reportlab==4.0.4
Pillow==10.0.0
PyMuPDF==1.23.5
